Vision revamp

+++ C-Engine/src/main/java/main/ability/PassiveAbilityObj.java
@@ -65,7 +65,7 @@ public class PassiveAbilityObj extends AbilityObj implements Attachment {
     }
 
     @Override
-    public int getDuration() {
+    public double getDuration() {
         return dur;
     }
